backend/requirements-prod.txt
# ── Syntora Backend — Production (no ML/Whisper) ──────────────────────────────
# Use this file for free-tier Railway / Render / Fly.io deployments.
# Rule-based transliteration still works perfectly without torch/transformers.
# For full neural engines, use requirements.txt on a machine with ≥2 GB RAM.

fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
sse-starlette>=1.6.5
